+++ src/components/PasswordReset/actions.js
@@ -18,8 +18,7 @@ export const resetPasswordFailure = errors => ({
 const sendEmailToUser = data => (dispatch) => {
   dispatch(resetPassword());
   axios
-    // .post('https://ah-codeofduty-staging.herokuapp.com/api/users/forgot-password/', data)
-    .post('http://127.0.0.1:8000/api/users/forgot-password/', data)
+    .post('https://ah-codeofduty-staging.herokuapp.com/api/users/forgot-password/', data)
     .then(response => dispatch(resetPasswordSuccess(response.data.message)))
     .catch((err) => {
       dispatch(resetPasswordFailure(err.response.data.error[0]));
