+++ src/components/PasswordResetForm/passwordresetform.test.js
@@ -1,6 +1,8 @@
 import expect from 'expect';
 import * as actions from './actions';
 import * as types from './constants';
+import * as action from '../PasswordReset/actions';
+import * as type from '../PasswordReset/constants';
 
 describe('Reset saved password i database', () => {
   it('start process of resetting password', () => {
@@ -23,5 +25,29 @@ describe('Reset saved password i database', () => {
       type: types.RESET_PASSWORD_ACTION_SUCCESS,
     };
     expect(actions.resetPasswordActionSuccess()).toEqual(expectedAction);
+  });
+
+  // ........................
+
+  it('start process of resetting password', () => {
+    const email = 'example@gmail.com';
+    const expectedAction = {
+      type: type.RESET_PASSWORD,
+    };
+    expect(action.resetPassword(email)).toEqual(expectedAction);
+  });
+
+  it('forgot password failure', () => {
+    const expectedAction = {
+      type: type.RESET_PASSWORD_FAILURE,
+    };
+    expect(action.resetPasswordFailure()).toEqual(expectedAction);
+  });
+
+  it('forgot password sucess', () => {
+    const expectedAction = {
+      type: type.RESET_PASSWORD_SUCCESS,
+    };
+    expect(action.resetPasswordSuccess()).toEqual(expectedAction);
   });
 });
